+++ src/Components/Input.test.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
+import { WebApps } from '../Context/index';
 
 import Input from './Input';
 
 test('Input Component Renders', () => {
-    render(<Input data-testid="webapps-Input-component" />);
+    render(<WebApps><Input data-testid="webapps-Input-component" /></WebApps>);
 
     expect(screen.getByTestId('webapps-Input-component')).toBeDefined();
 });